fix: fix nic settings

+++ onpre/scripts/nic.bash
@@ -2,52 +2,56 @@
 
 set -euo pipefail
 
-: "$VPN_NIC" "$VPN_DOMAIN"
-export PATH="/usr/local/bin:$PATH"
-
-checknic() {
-  dhclient "$VPN_NIC"
-  MY_IP=$(ip addr show enp2s0 | grep inet | awk '{print $2}' | awk -F/ '{print $1}' | head -n 1)
-  MY_GW=$(ip route | grep default | awk '{print $3}')
-  MY_TOP_IP=$(echo "$MY_IP" | awk -F. '{print $1".0.0.0/8"}')
-  VPN_IP=$(ip addr show vpn_vpnnic | grep inet | awk '{print $2}' | awk -F/ '{print $1}' | head -n 1)
-  VPN_GW=$(echo "$VPN_IP" | awk -F. '{print $1"."$2"."$3".1"}')
-  VPN_HOST=$(nslookup "$VPN_DOMAIN" | grep Address | awk '{print $2}' | tail -n 1)
-  echo "My IP: $MY_IP"
-  echo "My GW: $MY_GW"
-  echo "VPN DOMAIN: $VPN_DOMAIN"
-  echo "VPN HOST: $VPN_HOST"
-  echo "VPN IP: $VPN_IP"
-  echo "VPN GW: $VPN_GW"
-
-  ip route add "$VPN_HOST" via "$MY_GW"
-  ip route add "$MY_TOP_IP" via "$MY_GW"
-  ip route del default
-  ip route add default via "$VPN_GW"
-  echo "nameserver $VPN_GW" >/etc/resolv.conf
-}
+: "$VPN_DOMAIN"
 
 cat <<EOM
 
-## Copy the vpn_vpnnic interface
+## Cofigure the vpn_vpnnic interface
 EOM
 
-cat ./scripts/interfaces >/etc/network/interfaces.d/vpn_vpnnic
+rm -f /etc/network/interfaces.d/vpn_vpnnic
+
+if (! grep 'up ip route add' /etc/network/interfaces); then
+  VPN_IP="$(nslookup "$VPN_DOMAIN" | grep Address | awk '{print $2}' | tail -n 1)"
+  MY_IP="$(ip addr show enp2s0 | grep inet | awk '{print $2}' | awk -F/ '{print $1}' | head -n 1)"
+  sed -i "/iface vmbr0 inet static/a \\\tup ip route add $VPN_IP via $MY_IP" /etc/network/interfaces
+  sed -i "/iface vmbr0 inet static/a \\\tup ip route add 192.168.133.0/24 via 172.16.168.1" /etc/network/interfaces
+fi
+
+if (! grep '# vpn_vpnnic settings' /etc/network/interfaces); then
+  cat ./scripts/interfaces >> /etc/network/interfaces
+  MY_IP="$(ip addr show enp2s0 | grep inet | awk '{print $2}' | awk -F/ '{print $1}' | head -n 1)"
+  sed -i "/iface vmbr1 inet static/a \\\tup ip route add default via $MY_IP" /etc/network/interfaces
+fi
+
+ifdown vmbr0
+ifup vmbr0
+ifdown vmbr1
+ifup vmbr1
+
+systemctl restart networking
 
 systemctl restart vpnclient
+
+dhclient
 
 cat <<EOM
 
 ## Try connecting to the VPN
 EOM
 
-while true; do
+for i in {1..10}; do
   if ip link show vpn_vpnnic &>/dev/null; then
-    checknic
     wget -q -O - httpbin.org/ip
     break
   fi
   sleep 1
+  echo "Waiting for VPN connection... $i/10"
 done
+
+if ! ip link show vpn_vpnnic &>/dev/null; then
+  echo "Failed to establish VPN connection"
+  exit 1
+fi
 
 systemctl stop vpnclient
